[F005][F006] Moved around functions for more clarity in what each file does
